requirements.txt
aiohttp
SQLAlchemy==1.4.48
asyncpg
python-dotenv
more_itertools
